fixed date formatting issue and exported format date function to util file everywhere it is needed in app
